Added function in mainwindow for comboBox creation from parm file,  speciesSelection "deleteAll" functionality
Created a static function in mainwindow to allow dynamic comboBox construction in program (consider moving to GPSB if location makes more sense for code reuse).
addDynamComboBox function allows for the additional functionality of parms file instructions for comboBoxes that include "\n" to denote separate options inline.

Additionaly, a catch for speciesSelection's "deleteAll" functionality was added, along with " " (blank) being the default. The next species in the comboBox could easily be changed to the default if this makes more sense for usability.