+ parent integration to GPSB warning window, default parmMap result
When a window is passed into the GPSB warning window constructor, it will now center itself on the parent.  If no parent is passed into the GPSB window it defaults to the center of the screen.

When a Main Section Title is looked up in the parmMap, a default of -1 is passed in to ensure that any result >= 0 works.