+ dynamicArray, answer & parms Forms improvements, parseForm()
Suppose can now read the dynamicArrays within answerForms and parmsForms, and also applies them correctly.

The answerForm and parmsForms have been greatly improved, with answerForm getting the most attention.

The acceptedInput QVector is now built from the highest field number, insuring the correct amount of vector slots as well as the correct length and behavior of the answerStrings.

A dedicated function called parseForm was created for the separate answerForm and parmsForm. The methodology was also improved.

The accept() code's handling of parms and answer Forms was improved and a default answerForm was added.

Suppose now handles blank columns as well as extra words in answer and parms Forms.